Fix instanceof crash on IE
According to the MSDN documentation [1], in the Javascript engine of
Internet Explorer, the class passed into the instanceof method must be
defined or the instanceof method may crash.

This commit introduces a wrapper around instanceof that makes sure that we
never call `something instanceof null` or `something instanceof undefined`
thus avoiding this source of crashes.

[1] https://msdn.microsoft.com/en-us/library/zh0zb36z(v=vs.94).aspx#Parameters